refactor(ui): relocate OpenArtistRefInline to shared ui/ layer, breaking the album-artist cycle
