+++ src/features/album/components/AlbumCard.tsx
@@ -15,7 +15,7 @@ import type { CoverPrefetchPriority } from '@/cover/types';
 import { COVER_DENSE_GRID_MIN_CELL_CSS_PX } from '@/cover/layoutSizes';
 import { resolveCoverDisplayTier } from '@/cover/tiers';
 import { acquireUrl } from '@/cover';
-import { OpenArtistRefInline } from '@/features/artist';
+import { OpenArtistRefInline } from '@/ui/OpenArtistRefInline';
 import { fetchAlbumTracks, playAlbum, playAlbumShuffled } from '@/features/playback/utils/playback/playAlbum';
 import { useLongPressAction } from '@/lib/hooks/useLongPressAction';
 import { LongPressWaveOverlay } from '@/ui/LongPressWaveOverlay';
