+++ src/ui/OpenArtistRefInline.tsx
